Fix page loading issue and enhance project structure.
- Added missing loading screen to all pages to fix content visibility issue.
- Refactored monolithic script.js into smaller, more manageable modules (ui.js, cart.js, product.js).
- Moved hardcoded product data into a separate products.json file.
- Added "View Details" buttons to product cards on the index page.
- Added a confirmation dialog when clearing the cart.
