+++ index.html
@@ -89,6 +89,7 @@
                     <div class="product-image">
                         <img src="https://images.pexels.com/photos/2582937/pexels-photo-2582937.jpeg?auto=compress&cs=tinysrgb&w=400" alt="Gaming PC">
                         <div class="product-overlay">
+                            <button class="btn btn-small view-details" data-product="gaming-pc-ultimate">View Details</button>
                             <button class="btn btn-small add-to-cart" data-product="gaming-pc-ultimate">Add to Cart</button>
                         </div>
                     </div>
@@ -102,6 +103,7 @@
                     <div class="product-image">
                         <img src="https://images.pexels.com/photos/1029757/pexels-photo-1029757.jpeg?auto=compress&cs=tinysrgb&w=400" alt="Gaming Monitor">
                         <div class="product-overlay">
+                            <button class="btn btn-small view-details" data-product="monitor-4k">View Details</button>
                             <button class="btn btn-small add-to-cart" data-product="monitor-4k">Add to Cart</button>
                         </div>
                     </div>
@@ -115,6 +117,7 @@
                     <div class="product-image">
                         <img src="https://images.pexels.com/photos/2115257/pexels-photo-2115257.jpeg?auto=compress&cs=tinysrgb&w=400" alt="Gaming Keyboard">
                         <div class="product-overlay">
+                            <button class="btn btn-small view-details" data-product="keyboard-mechanical">View Details</button>
                             <button class="btn btn-small add-to-cart" data-product="keyboard-mechanical">Add to Cart</button>
                         </div>
                     </div>
@@ -128,6 +131,7 @@
                     <div class="product-image">
                         <img src="https://images.pexels.com/photos/3394650/pexels-photo-3394650.jpeg?auto=compress&cs=tinysrgb&w=400" alt="Gaming Headset">
                         <div class="product-overlay">
+                            <button class="btn btn-small view-details" data-product="headset-pro">View Details</button>
                             <button class="btn btn-small add-to-cart" data-product="headset-pro">Add to Cart</button>
                         </div>
                     </div>
@@ -245,6 +249,6 @@
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.13.0/gsap.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.13.0/ScrollTrigger.min.js"></script>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
 </body>
 </html>
